Unify naming to "Manager" in favor of "Controller" and change project code to reflect this
