+++ DNN Platform/Library/DotNetNuke.Library.csproj
@@ -220,7 +220,7 @@
     <Compile Include="Entities\Modules\Settings\ParameterAttributeBase.cs" />
     <Compile Include="Entities\Modules\Settings\ParameterMapping.cs" />
     <Compile Include="Entities\Modules\Settings\PortalSettingAttribute.cs" />
-    <Compile Include="Entities\Modules\Settings\SerializationController.cs" />
+    <Compile Include="Entities\Modules\Settings\SerializationManager.cs" />
     <Compile Include="Entities\Modules\Settings\SettingsRepository.cs" />
     <Compile Include="Entities\Modules\Settings\TabModuleSettingAttribute.cs" />
     <Compile Include="Entities\Portals\IPortalSettingHandlers.cs" />
